Landing functional
page navigation added
