+++ package.json
@@ -10,15 +10,20 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~3.2.1",
-    "expo": "~51.0.14",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.10.0",
+    "@react-navigation/stack": "^6.4.0",
+    "expo": "~51.0.19",
     "expo-linear-gradient": "~13.0.2",
     "expo-status-bar": "~1.12.1",
     "firebase": "^10.12.2",
     "galio-framework": "^0.8.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.74.2",
+    "react-native": "^0.74.3",
     "react-native-linear-gradient": "^2.8.3",
+    "react-native-safe-area-context": "^4.10.1",
+    "react-native-screens": "^3.31.1",
     "react-native-web": "~0.19.10"
   },
   "devDependencies": {
